+++ src/main/java/com/team1816/lib/motion/profiles/MotionProfile.java
@@ -1,4 +1,4 @@
-package com.team1816.lib.math.motion.profiles;
+package com.team1816.lib.motion.profiles;
 
 public abstract class MotionProfile {
