+++ src/main/java/com/team1816/lib/motion/profiles/SinusoidalMotionProfile.java
@@ -1,4 +1,4 @@
-package com.team1816.lib.math.motion.profiles;
+package com.team1816.lib.motion.profiles;
 
 /**
  * This class will construct a sinusoidal motion profile.
